Add files via upload
This program allows someone to search through and display sorted player data on a team. The playerData.txt file is data from a water polo team that includes the names, heights, GPAs, and more statistics about each player. A prompt is shown to the user and allows them to search through the top K of a given statistic on the team. There are three different methods that the program uses to sort: quicksort (recursive), insertion sort (iterative), and a third method that is another version of the iterative sort.